requirements.txt
yfinance>=0.2.31
pandas>=2.0.0
numpy>=1.24.0
backtesting>=0.3.3
PyYAML>=6.0
matplotlib>=3.7.0
ta>=0.10.2
click>=8.1.0
tabulate>=0.9.0
